# [PERF] Improve resource management in BenchmarkFromHuggingFace