## Summary
Split `BenchmarkFromHuggingFace` into two sub-benchmarks for clearer performance analysis:
- **CreationOnly**: Measures tokenizer creation time excluding `Close()` overhead
- **FullLifecycle**: Measures complete lifecycle including `Close()` overhead

## Implementation Details
- Modified `BenchmarkFromHuggingFace` to use sub-benchmarks with `b.Run()`
- Added `b.ReportAllocs()` to track memory allocation patterns
- Added `b.ResetTimer()` after warmup to exclude warmup time from measurements
- CreationOnly collects tokenizers in a slice and closes them after `b.StopTimer()` to avoid per-iteration timer overhead
- FullLifecycle preserves original behavior measuring total time including cleanup

## Benchmark Results
With `benchtime=1s`:
```
BenchmarkFromHuggingFace/CreationOnly-16   	      64	  18470865 ns/op
BenchmarkFromHuggingFace/FullLifecycle-16  	      58	  19838025 ns/op
```

The `Close()` overhead is approximately **1.37ms per operation** (19.84ms - 18.47ms).

Closes #54